Addin auth to file controller

+++ fileshare/Controllers/FileController.cs
@@ -5,9 +5,14 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using FileShare.Services.Interfaces;
 using FileShare.Contracts;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authentication.Cookies;
+using System.Security.Claims;
 
 namespace FileShare.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("[controller]")]
     public class FileController : ControllerBase
@@ -23,25 +28,32 @@ namespace FileShare.Controllers
         }
 
         [HttpGet]
-        public async Task<IEnumerable<FileDto>> Get(Guid roomId)
+        public async Task<IActionResult> Get(Guid roomId)
         {
-            return await _service.GetFiles(roomId);
+            if (roomId == Guid.Empty || !User.HasClaim(ClaimTypes.Name, roomId.ToString()))
+                return BadRequest("User does not have appropriate permissions");
+            return Ok(await _service.GetFiles(roomId));
         }
 
         [HttpGet("{id}")]
-        public async Task<FileDto> GetById(Guid id)
+        public async Task<IActionResult> GetById(Guid id)
         {
             if (id == Guid.Empty)
-                throw new Exception("No ID provided");          
-            return await _service.GetFileInfo(id);
+                return BadRequest("File ID is invalid");
+            var result = await _service.GetFileInfo(id);
+            if (!User.HasClaim(ClaimTypes.Name, result.RoomId.ToString()))
+                return BadRequest("User does not have appropriate permissions");
+            return Ok(result);
         }
 
         [HttpGet("{id}/download")]
         public async Task<IActionResult> Download(Guid id)
         {
             if (id == Guid.Empty)
-                throw new Exception("No file ID provided");          
+                return BadRequest("File ID is invalid");
             var fileInfo = await _service.GetFileInfo(id);
+            if (!User.HasClaim(ClaimTypes.Name, fileInfo.RoomId.ToString()))
+                return BadRequest("User does not have appropriate permissions"); 
             var results = await _service.GetFileDownload(fileInfo.RoomId, fileInfo.Name);
             return new FileContentResult(results.ToArray(), "application/octet-stream")
             {
@@ -53,14 +65,21 @@ namespace FileShare.Controllers
         public async Task<IActionResult> Create([FromForm] CreateFileDto dto)
         {
             if (dto == null)
-                throw new Exception("No data provided");       
+                throw new Exception("No data provided");    
+            if (!User.HasClaim(ClaimTypes.Name, dto.RoomId.ToString()))
+                return BadRequest("User does not have appropriate permissions");   
             var fileId = await _service.CreateFile(dto);
             return CreatedAtAction(nameof(Create), new { id = fileId });
         }
 
         [HttpDelete]
         public async Task<IActionResult> Delete(Guid id)
-        {     
+        {   
+            if (id == Guid.Empty)
+                return BadRequest("File ID is invalid");
+            var fileInfo = await _service.GetFileInfo(id);
+            if (!User.HasClaim(ClaimTypes.Name, fileInfo.RoomId.ToString()))
+                return BadRequest("User does not have appropriate permissions");
             await _service.DeleteFile(id);
             return NoContent();
         }
